requirements.txt
appnope==0.1.4
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==2.1.1
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
idna==3.10
ipykernel==6.15.2
ipython==8.31.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.2
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
Pygments==2.18.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
requests==2.28.1
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
tzdata==2024.2
urllib3==1.26.20
wcwidth==0.2.13
pandas==2.2.1
#matplotlib==3.7.1
#seaborn==0.12.2
#plotly==5.9.0
#polyline==2.0.0
#shapely==2.0.1
#geopandas==0.13.2
